ai_uml/requirements.txt
# Core dependencies
matplotlib
Pillow
Wand
svgwrite
defusedxml
lxml
transformers
matplotlib
diffusers

# For SVG to PNG conversion
cairosvg

# For model training and evaluation
torch>=1.10.0
transformers>=4.20.0
tqdm
umap-learn
numpy

# For visualization
matplotlib

torchvision
tqdm
accelerate
umap-learn
matplotlib
plotly
seaborn
